move 2pF optimisation functions to new lib optimisation file
